chore: propagate the profiler rename after rebasing onto PR 1
Rename detection moved this branch's files to their new homes during the rebase, but it
does not rewrite what is inside them. These are the references that still named the old
module:

  nemo_datasets_plugin.*           -> nemo_profiler_plugin.*
  ...profiler.file_source          -> nemo_profiler_plugin.source
  nemo_platform_plugin.files.dataset_profile -> ...files.profile

The load-bearing one is `_PROFILE_TASK_COMMAND` in `profile_job.py`: it is the module the
job actually execs, so a stale spelling there fails at run time rather than at import, and
no test that stubs the submission would have caught it.

`cpu-tasks` also carried both spellings after the rebase -- PR 1 now adds the plugin under
its new name, so this drops the duplicate this branch was contributing.

The OpenAPI spec and the generated SDK are unaffected: their names come from the
`DatasetProfile` schema, which did not change, and `make refresh-openapi` reports no drift.
